MT#55897 Fix links in the `README.md` / `overview.md`
It's been noticed that quick access links in the general `overview.md`
of the RTD are broken, they simply don't see needed files, since
these links were implemented for the github usage initially.

Because the `overview.md` is just a symlink to the `README.md` (used
by the github), it's impossible to have own links (to the github files)
for the `README.md` and own links for the `overview.md` in the RTD.

Keeping separate files for RTD and github's README would be bad idea,
since then we have to manually keep them synced.

Instead, this commit proposes the solution like this:
- have same quick access links for the RTD and the Github
- by clicking quick access links in the github, one will be sent
  to the RTD article.

It's a compromise solution, which gives us working access by both
github and RTD.

Then, in order to access github doc files, one has to enter the
`docs/` folder and select a needed one.

Change-Id: Ic58437dccbdccd8289e8282fcf8e91c71057e366
(cherry picked from commit c3893c214b17c2863e598545076b2a3ae580b3ef)